requirements.txt
streamlit==0.78.0
pandas==1.2.3
Pillow==8.2.0
scikit_learn==0.24.1
